Feat: Dialog  confirmar eliminacion usuario

+++ src/app/user/user-edit/user-edit.component.ts
@@ -4,6 +4,8 @@ import { UserService } from '../services/user.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { switchMap } from 'rxjs';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmComponent } from '../components/confirm/confirm.component';
 
 @Component({
   selector: 'app-user-edit',
@@ -20,7 +22,7 @@ export class UserEditComponent implements OnInit {
     email: ''
   }
 
-  constructor( private userService: UserService, private activatedRoute: ActivatedRoute, private router: Router, private snackBar: MatSnackBar ) { }
+  constructor( private userService: UserService, private activatedRoute: ActivatedRoute, private router: Router, private snackBar: MatSnackBar, private dialog: MatDialog ) { }
 
   ngOnInit(): void {
     this.activatedRoute.params.pipe( switchMap( ({ id }) => this.userService.getUserDetail(id)) ).subscribe( user => this.user = user );
@@ -41,7 +43,16 @@ export class UserEditComponent implements OnInit {
   }
 
   delete(){
-    this.userService.deleteUser( this.user.id!.toString() ).subscribe( resp => { this.router.navigate(['/users']), this.showSnackBar('Usuario eliminado con exito!') });
+    const dialog = this.dialog.open( ConfirmComponent, {
+      width: '300px',
+      data: this.user
+    } );
+
+    dialog.afterClosed().subscribe( (result ) => {
+      if (result) {
+        this.userService.deleteUser( this.user.id!.toString() ).subscribe( resp => { this.router.navigate(['/users']), this.showSnackBar('Usuario eliminado con exito!') });
+      }
+    } );
   }
 
   showSnackBar(message: string){
